Extract Docker Compose integration into its own module
container.rs had grown to ~3400 lines mixing two concerns: the container
runtime/registry mechanics (podman/docker, image push/pin, manifests) and the
Docker Compose integration (resolving the CLI, driving compose, parsing configs,
pinning service images).

Move everything that reads or manipulates a docker-compose.yaml into a new
`compose` module — ComposeCli/DockerCompose, config/ps parsing, image
availability + architecture validation, and the config-image staging — together
with their tests. container.rs keeps the runtime/image primitives, exposing the
few shared helpers (image-ref parsing, output_with_timeout) as pub(crate).

container.rs: ~3400 -> ~1400 lines; compose.rs: ~2000 lines. Pure move: 425
tests pass unchanged, clippy clean.
