+++ src/container.rs
@@ -1,6 +1,5 @@
-use std::collections::{BTreeMap, HashSet};
 use std::io::{Read, Write};
-use std::path::{Path, PathBuf};
+use std::path::Path;
 use std::process::{Command, Output, Stdio};
 use std::thread;
 use std::time::{Duration, Instant};
@@ -8,11 +7,10 @@ use std::time::{Duration, Instant};
 use chrono::Utc;
 use color_eyre::{
     Section, SectionExt,
-    eyre::{Context, Result, bail, eyre},
+    eyre::{Context, Result, eyre},
 };
 use tokio::process::Child;
 
-use crate::config::ComposeConfig;
 use crate::settings::Settings;
 
 /// Wall-clock budget for each synchronous docker/podman call made while
@@ -22,7 +20,7 @@ use crate::settings::Settings;
 /// whole CLI — forever, with neither `--timeout` nor ctrl+c able to interrupt a
 /// blocking `Command`. These calls are normally sub-second; the generous bound
 /// only exists to convert an indefinite hang into a clear error.
-const DISCOVERY_COMMAND_TIMEOUT: Duration = Duration::from_secs(60);
+pub(crate) const DISCOVERY_COMMAND_TIMEOUT: Duration = Duration::from_secs(60);
 
 /// Run a command to completion with a wall-clock timeout, killing it (and
 /// returning an error) if it overruns. Reader threads drain stdout/stderr so a
@@ -37,7 +35,7 @@ const DISCOVERY_COMMAND_TIMEOUT: Duration = Duration::from_secs(60);
 /// `exec test -d`, `ps`) whose only child is the client itself: killing it
 /// closes the pipes (so the reader threads finish) and the work we were waiting
 /// on ends. The daemon-side operation is intentionally not ours to kill.
-fn output_with_timeout(mut cmd: Command, timeout: Duration) -> Result<Output> {
+pub(crate) fn output_with_timeout(mut cmd: Command, timeout: Duration) -> Result<Output> {
     cmd.stdin(Stdio::null())
         .stdout(Stdio::piped())
         .stderr(Stdio::piped());
@@ -143,17 +141,6 @@ impl Drop for ProcessGroupChild {
             let _ = child.try_wait();
         }
     }
-}
-
-/// Build the `-f` flags for `compose` subcommands: always
-/// `-f docker-compose.yaml`, plus `-f <overlay>` if an overlay was provided.
-fn compose_file_args(overlay: Option<&Path>) -> Vec<String> {
-    let mut args = vec!["-f".to_string(), "docker-compose.yaml".to_string()];
-    if let Some(path) = overlay {
-        args.push("-f".to_string());
-        args.push(path.display().to_string());
-    }
-    args
 }
 
 /// Trait representing a container runtime (podman or docker).
@@ -481,621 +468,6 @@ fn stderr_indicates_missing_source(stderr: &str) -> bool {
         || s.contains("could not be found on container")
 }
 
-/// The standalone Docker Compose v2 binary name.
-const DOCKER_COMPOSE: &str = "docker-compose";
-
-/// How Docker Compose v2 is invoked on this machine.
-///
-/// Compose v2 ships two ways — the standalone `docker-compose` binary and the
-/// `docker compose` CLI plugin — and snouty drives whichever it finds. Modeling
-/// the two as an enum (rather than a program plus a free-form argument prefix)
-/// makes the wrong combinations unrepresentable: each variant fixes its own
-/// invocation. The wrapped path is always absolute so the command survives the
-/// `env_clear()` in the hermetic render, where `PATH` is gone.
-#[derive(Clone, Debug)]
-enum ComposeCli {
-    /// The standalone `docker-compose` binary; the path is `docker-compose`.
-    Standalone(PathBuf),
-    /// The `docker compose` CLI plugin; the path is the `docker` binary.
-    Plugin(PathBuf),
-}
-
-impl ComposeCli {
-    /// Resolve a Docker Compose v2 invocation and its version banner, with a
-    /// clear error when none is available. The banner is captured during the
-    /// v2 check so callers that want to display it need not spawn `version`
-    /// again.
-    ///
-    /// Prefers the standalone `docker-compose` binary when it is present and
-    /// genuinely v2 (the historical contract). Otherwise falls back to the
-    /// `docker compose` CLI plugin — but only when `docker` is real Docker,
-    /// never podman in disguise, whose compose provider may not implement the
-    /// v2 features snouty relies on.
-    fn resolve() -> Result<(ComposeCli, String)> {
-        // 1. Standalone docker-compose, when it's really v2.
-        if let Ok(program) = which::which(DOCKER_COMPOSE) {
-            let candidate = ComposeCli::Standalone(program);
-            match candidate.version() {
-                Ok(version) => return Ok((candidate, version)),
-                // Present but not v2 (likely Compose v1). Prefer the docker
-                // plugin if it's usable; only surface the v1 error if not.
-                Err(standalone_err) => {
-                    return Self::plugin().or(Err(standalone_err));
-                }
-            }
-        }
-
-        // 2. The `docker compose` CLI plugin. Preserve the plugin's own error
-        // (podman-in-disguise, a v1 plugin, or docker genuinely absent) as the
-        // cause, rather than collapsing every case to a generic "not found".
-        Self::plugin().map_err(|plugin_err| {
-            eyre!(
-                "snouty requires Docker Compose v2, but neither the `docker-compose` binary nor the `docker compose` CLI plugin is usable"
-            )
-            .with_section(move || format!("{plugin_err:#}").header("docker compose:"))
-            .with_suggestion(|| {
-                "install Docker Compose v2: https://docs.docker.com/compose/install/"
-            })
-        })
-    }
-
-    /// The `docker compose` plugin as a v2 invocation, paired with its version
-    /// banner. `Err` when docker is absent, is podman in disguise, or its
-    /// compose plugin isn't v2 — callers treat any of these as "no usable
-    /// plugin".
-    fn plugin() -> Result<(ComposeCli, String)> {
-        let program = which::which("docker").wrap_err("`docker` not found on PATH")?;
-        // podman-in-disguise routes `docker compose` to a provider that may not
-        // implement Compose v2; never trust it as a v2 source.
-        if is_podman_in_disguise("docker") {
-            bail!("`docker` is podman in disguise; its `compose` is not Docker Compose v2");
-        }
-        let candidate = ComposeCli::Plugin(program);
-        let version = candidate.version()?;
-        Ok((candidate, version))
-    }
-
-    /// The invocation as a user would type it, for user-facing hints.
-    fn display(&self) -> &'static str {
-        match self {
-            ComposeCli::Standalone(_) => "docker-compose",
-            ComposeCli::Plugin(_) => "docker compose",
-        }
-    }
-
-    /// A fresh [`Command`] for this invocation, positioned so callers append
-    /// only the compose subcommand and its arguments. The program (and the
-    /// plugin's leading `compose`) are fixed by the variant and can't be
-    /// clobbered by later `args`.
-    fn command(&self) -> Command {
-        match self {
-            ComposeCli::Standalone(program) => Command::new(program),
-            ComposeCli::Plugin(program) => {
-                let mut cmd = Command::new(program);
-                cmd.arg("compose");
-                cmd
-            }
-        }
-    }
-
-    /// Return the version banner, erroring when the command fails to run or
-    /// reports a non-v2 (v1) banner.
-    fn version(&self) -> Result<String> {
-        let name = self.display();
-        let mut cmd = self.command();
-        cmd.arg("version");
-        match cmd.output() {
-            Ok(o) if o.status.success() => {
-                let version = String::from_utf8_lossy(&o.stdout).trim().to_string();
-                if is_compose_v2_version(&version) {
-                    Ok(version)
-                } else {
-                    Err(eyre!("`{name}` is not Docker Compose v2"))
-                        .with_section(move || version.header(format!("{name} version:")))
-                        .with_suggestion(|| {
-                            "install Docker Compose v2: https://docs.docker.com/compose/install/"
-                        })
-                }
-            }
-            Ok(o) => {
-                let stderr = String::from_utf8_lossy(&o.stderr);
-                Err(eyre!("`{name} version` failed"))
-                    .with_section(move || stderr.trim().to_string().header("Stderr:"))
-            }
-            Err(e) => Err(eyre!("failed to run `{name} version`: {e}")),
-        }
-    }
-}
-
-/// Whether a `compose version` banner identifies Docker Compose v2. Both the
-/// standalone binary and the CLI plugin print "Docker Compose version ..." for
-/// v2; Compose v1's "docker-compose version ..." (hyphenated) does not match.
-fn is_compose_v2_version(output: &str) -> bool {
-    output.to_lowercase().contains("docker compose")
-}
-
-/// Drives Docker Compose v2, independent of which runtime built or pushed the
-/// images. Compose is invoked through whichever v2 form is available (see
-/// [`ComposeCli`]); `docker_host`, when set, points it at a specific engine
-/// (e.g. podman's API socket); when `None`, compose uses its default (the
-/// Docker daemon, or an explicit `DOCKER_HOST` inherited from the environment).
-///
-/// Image operations (tag, push, pin) are not compose's concern — the methods
-/// that need a container engine (e.g. [`pin_images`](Self::pin_images)) take a
-/// [`ContainerRuntime`] argument rather than this type owning one.
-pub struct DockerCompose {
-    cli: ComposeCli,
-    docker_host: Option<String>,
-}
-
-impl DockerCompose {
-    /// Resolve Docker Compose v2 and wire it to `rt`'s container engine.
-    ///
-    /// An explicit `DOCKER_HOST` already set in the environment is always
-    /// respected; otherwise, for a podman runtime, compose is pointed at
-    /// podman's API socket so podman backs Compose.
-    pub fn resolve(rt: &dyn ContainerRuntime) -> Result<DockerCompose> {
-        let (cli, _version) = ComposeCli::resolve()?;
-        let docker_host = if std::env::var_os("DOCKER_HOST").is_some() {
-            None
-        } else {
-            rt.engine_docker_host()?
-        };
-        Ok(DockerCompose { cli, docker_host })
-    }
-
-    /// Locate a usable Docker Compose v2 without wiring it to an engine, for
-    /// diagnostics and availability checks (`snouty doctor`, tests). Returns the
-    /// command name (`docker-compose` / `docker compose`) and version banner.
-    pub fn probe() -> Result<(&'static str, String)> {
-        let (cli, version) = ComposeCli::resolve()?;
-        Ok((cli.display(), version))
-    }
-
-    /// The command name snouty drives (`docker-compose` / `docker compose`),
-    /// for user-facing hints.
-    pub fn display(&self) -> &'static str {
-        self.cli.display()
-    }
-
-    /// The `DOCKER_HOST` compose is wired to, if any. Used to reproduce the
-    /// compose invocation in user-facing hints.
-    pub fn docker_host(&self) -> Option<&str> {
-        self.docker_host.as_deref()
-    }
-
-    /// Base compose command with engine wiring and working directory applied.
-    /// Callers append the compose subcommand and its arguments.
-    fn command(&self, config: &ComposeConfig) -> Command {
-        let mut cmd = self.cli.command();
-        cmd.current_dir(config.dir());
-        if let Some(host) = &self.docker_host {
-            cmd.env("DOCKER_HOST", host);
-        }
-        cmd
-    }
-
-    /// Async variant of [`command`](Self::command).
-    fn tokio_command(&self, config: &ComposeConfig) -> tokio::process::Command {
-        self.command(config).into()
-    }
-
-    /// Run `docker-compose config [extra_args]`, returning the resolved YAML
-    /// as a string.
-    fn config_yaml(
-        &self,
-        config: &ComposeConfig,
-        overlay: Option<&Path>,
-        extra_args: &[&str],
-    ) -> Result<String> {
-        // No COMPOSE_PROJECT_NAME override: the project name must resolve
-        // exactly as it does when the user runs `docker compose` in the
-        // config dir, because default build tags are derived from it.
-        let mut cmd = self.command(config);
-        cmd.args(compose_file_args(overlay));
-        cmd.arg("config");
-        cmd.args(extra_args);
-        let output = cmd
-            .output()
-            .wrap_err("failed to run 'docker-compose config'")?;
-
-        if !output.status.success() {
-            let stderr = String::from_utf8_lossy(&output.stderr);
-            let stdout = String::from_utf8_lossy(&output.stdout);
-            return Err(eyre!("'docker-compose config' failed"))
-                .with_section(move || stdout.trim().to_string().header("Stdout:"))
-                .with_section(move || stderr.trim().to_string().header("Stderr:"));
-        }
-
-        Ok(String::from_utf8_lossy(&output.stdout).into_owned())
-    }
-
-    /// Resolve and parse a compose config into structured contents.
-    pub fn contents(
-        &self,
-        config: &ComposeConfig,
-        overlay: Option<&Path>,
-    ) -> Result<ComposeContents> {
-        let yaml = self.config_yaml(config, overlay, &[])?;
-        parse_compose_config(&yaml)
-    }
-
-    /// Resolve the compose file to JSON using the normal (local) environment —
-    /// the same interpolation `snouty` sees when it runs compose on this machine.
-    pub fn config_json(&self, config: &ComposeConfig) -> Result<String> {
-        self.config_yaml(config, None, &["--format", "json"])
-    }
-
-    /// Resolve the compose file to JSON under a scrubbed process environment
-    /// that mimics the hermetic Antithesis environment: none of the user's shell
-    /// variables, so `${VAR}` interpolation resolves only from the config dir's
-    /// `.env` file, explicit env files, and inline defaults. The compose binary
-    /// was resolved before the environment is cleared, so no shell variables
-    /// need to be retained.
-    ///
-    /// Returns the raw output rather than a string: a required `${VAR:?}` with no
-    /// value makes compose abort (non-zero exit, empty stdout), which the caller
-    /// reads as a definite "won't resolve in Antithesis".
-    pub fn config_json_hermetic_env(&self, config: &ComposeConfig) -> Result<Output> {
-        // Reuse the normal command (binary + working directory), then clear the
-        // whole environment — including the DOCKER_HOST that command() sets.
-        // Those are all valid interpolation inputs Antithesis will not inherit,
-        // so scrubbing them is the point; only the binary and directory carry
-        // over. (env_clear() drops anything command() set via .env().)
-        let mut cmd = self.command(config);
-        cmd.env_clear();
-        cmd.args(compose_file_args(None));
-        cmd.args(["config", "--format", "json"]);
-        cmd.output()
-            .wrap_err("failed to run 'docker-compose config' for the environment check")
-    }
-
-    /// Canonicalized compose file for baking into the config image.
-    ///
-    /// `docker-compose config` itself does the canonicalization: anchors,
-    /// aliases, and merge keys are inlined, and the structure is normalized.
-    /// `--no-interpolate` keeps `${VAR}` references for the platform to
-    /// resolve in its own environment, and `--no-path-resolution` keeps
-    /// relative paths relative — both would otherwise be baked with values
-    /// from this machine.
-    fn canonical_contents(&self, config: &ComposeConfig) -> Result<String> {
-        self.config_yaml(config, None, &["--no-interpolate", "--no-path-resolution"])
-    }
-
-    /// Parse `docker-compose ps -a --format json` into the list of containers,
-    /// including stopped/exited ones so callers can flag stranded test
-    /// commands. Inspect [`ComposeContainer::stopped`] to tell them apart.
-    pub fn ps(
-        &self,
-        config: &ComposeConfig,
-        overlay: Option<&Path>,
-    ) -> Result<Vec<ComposeContainer>> {
-        let mut cmd = self.command(config);
-        cmd.args(compose_file_args(overlay));
-        cmd.args(["ps", "-a", "--format", "json"]);
-
-        let output = output_with_timeout(cmd, DISCOVERY_COMMAND_TIMEOUT)
-            .wrap_err("failed to run 'docker-compose ps'")?;
-
-        if !output.status.success() {
-            let stderr = String::from_utf8_lossy(&output.stderr);
-            return Err(eyre!("'docker-compose ps' failed"))
-                .with_section(move || stderr.trim().to_string().header("Stderr:"));
-        }
-
-        let stdout = String::from_utf8_lossy(&output.stdout);
-        parse_compose_ps(&stdout)
-    }
-
-    /// Run a command inside a running compose service container.
-    ///
-    /// Runs `docker-compose exec -T [--workdir workdir] {service} {cmd...}`.
-    /// The `-T` flag disables TTY allocation for non-interactive use. If
-    /// `workdir` is `Some`, sets the working directory inside the container.
-    /// Stdout and stderr are captured in the returned `Output`.
-    pub fn exec(
-        &self,
-        config: &ComposeConfig,
-        overlay: Option<&Path>,
-        service: &str,
-        workdir: Option<&str>,
-        env: &[(&str, &str)],
-        cmd: &[&str],
-    ) -> Result<std::process::Output> {
-        let mut command = self.command(config);
-        command.args(compose_file_args(overlay));
-        command.args(["exec", "-T"]);
-        for (k, v) in env {
-            command.args(["-e", &format!("{k}={v}")]);
-        }
-        if let Some(w) = workdir {
-            command.args(["--workdir", w]);
-        }
-        command.arg(service);
-        command.args(cmd);
-
-        command
-            .output()
-            .wrap_err("failed to run 'docker-compose exec'")
-    }
-
-    /// Spawn `docker-compose up --detach` and return the child process.
-    ///
-    /// stdout and stderr are inherited so progress is visible during pulls. The
-    /// caller awaits the child and checks its exit status. Uses
-    /// `process_group(0)` so the whole group can be killed on timeout.
-    pub fn up_detached(
-        &self,
-        config: &ComposeConfig,
-        overlay: Option<&Path>,
-    ) -> Result<ProcessGroupChild> {
-        let mut cmd = self.tokio_command(config);
-        cmd.args(compose_file_args(overlay));
-        cmd.args(["up", "--detach", "--no-build", "--pull=never"]);
-        cmd.stdin(std::process::Stdio::null());
-        cmd.stdout(std::process::Stdio::inherit());
-        cmd.stderr(std::process::Stdio::inherit());
-        cmd.process_group(0);
-
-        cmd.spawn()
-            .map(ProcessGroupChild::new)
-            .wrap_err("failed to start 'docker-compose up --detach'")
-    }
-
-    /// Spawn `docker-compose logs --follow` and return the child process.
-    ///
-    /// stdout and stderr are inherited so log output goes straight to the
-    /// terminal. stdin is null. The process exits when all containers stop.
-    pub fn logs_follow(
-        &self,
-        config: &ComposeConfig,
-        overlay: Option<&Path>,
-    ) -> Result<ProcessGroupChild> {
-        let mut cmd = self.tokio_command(config);
-        cmd.args(compose_file_args(overlay));
-        cmd.args(["logs", "--follow"]);
-        cmd.stdin(std::process::Stdio::null());
-        cmd.stdout(std::process::Stdio::inherit());
-        cmd.stderr(std::process::Stdio::inherit());
-        cmd.process_group(0);
-
-        cmd.spawn()
-            .map(ProcessGroupChild::new)
-            .wrap_err("failed to start 'docker-compose logs --follow'")
-    }
-
-    /// Run `docker-compose down` for cleanup. Best-effort, ignores errors.
-    pub fn down(&self, config: &ComposeConfig, overlay: Option<&Path>) {
-        let mut cmd = self.command(config);
-        cmd.args(compose_file_args(overlay));
-        cmd.args(["down", "--timeout", "0"]);
-        cmd.stdin(std::process::Stdio::null());
-        cmd.stdout(std::process::Stdio::null());
-        cmd.stderr(std::process::Stdio::null());
-        let _ = cmd.status();
-    }
-
-    /// Resolve every compose service image to a digest-pinned reference and
-    /// return the `docker-compose.yaml` contents canonicalized and rewritten
-    /// with those pins (`name:tag@sha256:...`).
-    ///
-    /// The local image store is the single source of truth for what a launch
-    /// runs — snouty never pulls. Every service must resolve to an image that
-    /// is present locally (built via its `build:` stanza, built or loaded out
-    /// of band, or previously pulled). Each image is then pinned to its local
-    /// digest in a registry confirmed to serve it ([`Self::find_remote_pin`]),
-    /// or — when no registry has it — tagged with the `registry` prefix and
-    /// pushed, so the platform always pulls exactly what was resolved here.
-    pub fn pin_images(
-        &self,
-        rt: &dyn ContainerRuntime,
-        config: &ComposeConfig,
-        registry: &str,
-    ) -> Result<String> {
-        let contents = self.contents(config, None)?;
-        with_config_image_escape_hatch(validate_images_are_available(rt, &contents))?;
-
-        let prefix = format!("{}/", registry.trim_end_matches('/'));
-
-        // Resolve each distinct image once: pin it from a registry that
-        // already serves the local digest, or schedule it for push.
-        let mut resolution: BTreeMap<&str, Option<String>> = BTreeMap::new();
-        for service in &contents.services {
-            let image = service.image.as_str();
-            if !resolution.contains_key(image) {
-                let pin = self.find_remote_pin(rt, image, &prefix)?;
-                if let Some(pinned_ref) = &pin {
-                    eprintln!("Image already in a registry, skipping push: {pinned_ref}");
-                }
-                resolution.insert(image, pin);
-            }
-        }
-
-        // service name -> final pinned reference (filled in for push targets
-        // after their pushes complete).
-        let mut pinned: BTreeMap<String, String> = BTreeMap::new();
-        // (service name, registry reference) for images we push ourselves.
-        let mut push_targets: Vec<(String, String)> = Vec::new();
-        let mut tagged: HashSet<&str> = HashSet::new();
-        for service in &contents.services {
-            let image = service.image.as_str();
-            if let Some(remote) = &resolution[image] {
-                pinned.insert(service.name.clone(), remote.clone());
-                continue;
-            }
-            let dest = if image.starts_with(&prefix) {
-                image.to_string()
-            } else {
-                format!("{prefix}{image}")
-            };
-            if dest != image && tagged.insert(image) {
-                rt.image_tag(image, &dest)?;
-            }
-            push_targets.push((service.name.clone(), dest));
-        }
-
-        // Arch-check and push each distinct image, pinning every service to
-        // its push digest (the push already reports the digest). The local
-        // architecture check applies exactly to the images whose local bytes
-        // we upload; remote pins were already amd64-verified.
-        let mut seen = HashSet::new();
-        let dests: Vec<&str> = push_targets
-            .iter()
-            .map(|(_, dest)| dest.as_str())
-            .filter(|dest| seen.insert(*dest))
-            .collect();
-        validate_image_architectures(rt, &dests)?;
-        let mut digests: BTreeMap<&str, String> = BTreeMap::new();
-        for dest in &dests {
-            eprintln!("Pushing image: {dest}");
-            let pinned_ref = rt.image_push(dest)?;
-            eprintln!("Image pushed: {pinned_ref}");
-            digests.insert(dest, pinned_ref);
-        }
-        for (name, dest) in &push_targets {
-            pinned.insert(name.clone(), digests[dest.as_str()].clone());
-        }
-
-        rewrite_compose_images(&self.canonical_contents(config)?, &pinned)
-    }
-
-    /// Find a registry that already serves `image`'s local bytes, returning
-    /// the digest-pinned reference to use, or `None` when the image must be
-    /// pushed.
-    ///
-    /// Candidate digests come from the local store's repo digests, for two
-    /// repositories: the image's own (e.g. `docker.io/library/redis` for
-    /// `redis:7`) and its `prefix`ed name from a previous snouty push. A
-    /// candidate counts only when the registry confirms it serves the digest
-    /// (a manifest-only round trip — never a pull or push) AND the platform
-    /// can run amd64 from it: a manifest list must offer an amd64 entry,
-    /// while a single manifest shares the local image's architecture, so the
-    /// local image must be amd64.
-    fn find_remote_pin(
-        &self,
-        rt: &dyn ContainerRuntime,
-        image: &str,
-        prefix: &str,
-    ) -> Result<Option<String>> {
-        let repo_digests = rt.image_repo_digests(image)?;
-        let tag = image_ref_tag(image);
-
-        let mut repos = vec![normalize_repo(image_repo(image))];
-        if !image.starts_with(prefix) {
-            repos.push(normalize_repo(image_repo(&format!("{prefix}{image}"))));
-        }
-
-        for repo in &repos {
-            // A pull typically records several digests per repo (the
-            // per-arch manifest and the manifest list) — try them all.
-            for digest in digests_for_repo(repo, &repo_digests) {
-                let amd64_ok = match rt.remote_manifest(&format!("{repo}@{digest}")) {
-                    RemoteManifest::NotFound => continue,
-                    RemoteManifest::List { has_amd64 } => has_amd64,
-                    RemoteManifest::Single => rt.image_architecture(image)? == "amd64",
-                };
-                if amd64_ok {
-                    return Ok(Some(format!("{repo}:{tag}@{digest}")));
-                }
-                // Served, but not runnable as amd64 — keep looking; the push
-                // path's local arch check produces the actionable error.
-            }
-        }
-        Ok(None)
-    }
-}
-
-/// Rewrite each service's `image:` field to its pinned digest reference.
-///
-/// Every service in the document must have an entry in `pinned` — the whole
-/// point of the rewrite is that the platform runs only digest-pinned images,
-/// so a service this function can't pin means pinning lost track of it
-/// somewhere upstream, which is a bug.
-fn rewrite_compose_images(yaml: &str, pinned: &BTreeMap<String, String>) -> Result<String> {
-    let mut doc: serde_yaml::Value =
-        serde_yaml::from_str(yaml).wrap_err("failed to parse docker-compose.yaml")?;
-    let services = doc
-        .get_mut("services")
-        .and_then(|s| s.as_mapping_mut())
-        .ok_or_else(|| eyre!("compose config has no services — this is a bug in snouty"))?;
-    for (name, svc) in services.iter_mut() {
-        let name = name
-            .as_str()
-            .ok_or_else(|| eyre!("compose config has a non-string service name: {name:?}"))?;
-        let pinned_ref = pinned.get(name).ok_or_else(|| {
-            eyre!("service '{name}' did not resolve to a pinned image reference — this is a bug in snouty")
-        })?;
-        let svc_map = svc
-            .as_mapping_mut()
-            .ok_or_else(|| eyre!("compose service '{name}' is not a mapping"))?;
-        svc_map.insert(
-            serde_yaml::Value::String("image".to_string()),
-            serde_yaml::Value::String(pinned_ref.clone()),
-        );
-    }
-    serde_yaml::to_string(&doc).wrap_err("failed to serialize pinned docker-compose.yaml")
-}
-
-/// Every compose service must carry an explicit `image:` reference; pinning
-/// (and the Antithesis platform) addresses services by image. In particular a
-/// `build:`-only service would otherwise silently run under a compose-generated
-/// default name that snouty never pushed.
-fn ensure_services_have_images(contents: &ComposeContents) -> Result<()> {
-    if contents.services_without_image.is_empty() {
-        return Ok(());
-    }
-    let mut err = eyre!("every compose service needs an explicit `image:` reference");
-    for name in &contents.services_without_image {
-        err = err.with_note(|| format!("service '{name}' has no `image:` field"));
-    }
-    Err(err.with_suggestion(|| {
-        "add an `image:` field to each service (for `build:` services, the tag the build produces)"
-    }))
-}
-
-/// Stage a copy of `config_dir` with `docker-compose.yaml` replaced by
-/// `pinned_yaml`, so the config image is built from the digest-pinned compose.
-/// The returned [`tempfile::TempDir`] must be kept alive until the image build
-/// completes.
-///
-/// Symlinks are recreated as-is (not dereferenced): a `docker build` context
-/// tars symlinks verbatim too, so the staged tree produces the same image
-/// content as building from `config_dir` directly.
-pub fn stage_pinned_config(config_dir: &Path, pinned_yaml: &str) -> Result<tempfile::TempDir> {
-    let staged = tempfile::tempdir().wrap_err("failed to create config staging directory")?;
-    copy_dir_recursive(config_dir, staged.path())?;
-    std::fs::write(staged.path().join("docker-compose.yaml"), pinned_yaml)
-        .wrap_err("failed to write pinned docker-compose.yaml")?;
-    Ok(staged)
-}
-
-/// Recursively copy the contents of `src` into `dst` (which must already exist).
-fn copy_dir_recursive(src: &Path, dst: &Path) -> Result<()> {
-    for entry in
-        std::fs::read_dir(src).wrap_err_with(|| format!("failed to read {}", src.display()))?
-    {
-        let entry = entry.wrap_err_with(|| format!("failed to read entry in {}", src.display()))?;
-        let file_type = entry.file_type().wrap_err("failed to read file type")?;
-        let from = entry.path();
-        let to = dst.join(entry.file_name());
-        if file_type.is_dir() {
-            std::fs::create_dir_all(&to)
-                .wrap_err_with(|| format!("failed to create {}", to.display()))?;
-            copy_dir_recursive(&from, &to)?;
-        } else if file_type.is_symlink() {
-            let target = std::fs::read_link(&from)
-                .wrap_err_with(|| format!("failed to read symlink {}", from.display()))?;
-            std::os::unix::fs::symlink(&target, &to)
-                .wrap_err_with(|| format!("failed to create symlink {}", to.display()))?;
-        } else {
-            std::fs::copy(&from, &to).wrap_err_with(|| {
-                format!("failed to copy {} to {}", from.display(), to.display())
-            })?;
-        }
-    }
-    Ok(())
-}
-
 #[derive(Clone)]
 pub struct PodmanRuntime {
     cmd: String,
@@ -1396,7 +768,7 @@ pub fn available_engines() -> Vec<Box<dyn ContainerRuntime>> {
 /// digest. The tag is kept as human-readable provenance — when inspecting an
 /// Antithesis run, the reference still shows which tag the digest came from.
 /// If the ref already carries a digest, it's replaced.
-fn pinned_image_ref(image_ref: &str, digest: &str) -> String {
+pub(crate) fn pinned_image_ref(image_ref: &str, digest: &str) -> String {
     match image_ref.rfind('@') {
         Some(at) => format!("{}@{}", &image_ref[..at], digest),
         None => format!("{image_ref}@{digest}"),
@@ -1465,7 +837,7 @@ fn classify_manifest_json(stdout: &[u8]) -> RemoteManifest {
 /// The repository part of an image reference: strips any `@digest` suffix and
 /// any `:tag` (a colon counts as a tag separator only after the last `/`;
 /// before it, it's a registry port).
-fn image_repo(image_ref: &str) -> &str {
+pub(crate) fn image_repo(image_ref: &str) -> &str {
     let no_digest = match image_ref.rfind('@') {
         Some(at) => &image_ref[..at],
         None => image_ref,
@@ -1483,7 +855,7 @@ fn image_repo(image_ref: &str) -> &str {
 }
 
 /// The tag of an image reference, or `latest` when untagged.
-fn image_ref_tag(image_ref: &str) -> &str {
+pub(crate) fn image_ref_tag(image_ref: &str) -> &str {
     let no_digest = match image_ref.rfind('@') {
         Some(at) => &image_ref[..at],
         None => image_ref,
@@ -1501,7 +873,7 @@ fn image_ref_tag(image_ref: &str) -> &str {
 /// `user/app` → `docker.io/user/app`, `index.docker.io/...` → `docker.io/...`.
 /// Repositories naming any other registry (first component with a dot, a
 /// port, or `localhost`) pass through unchanged.
-fn normalize_repo(repo: &str) -> String {
+pub(crate) fn normalize_repo(repo: &str) -> String {
     let (registry, rest) = match repo.split_once('/') {
         Some((first, rest))
             if first.contains('.') || first.contains(':') || first == "localhost" =>
@@ -1526,7 +898,7 @@ fn normalize_repo(repo: &str) -> String {
 /// entries, comparing normalized repository names. Multiple entries per repo
 /// are common — a pull records both the per-arch manifest digest and the
 /// manifest-list digest.
-fn digests_for_repo(repo: &str, repo_digests: &[String]) -> Vec<String> {
+pub(crate) fn digests_for_repo(repo: &str, repo_digests: &[String]) -> Vec<String> {
     let want = normalize_repo(repo);
     repo_digests
         .iter()
@@ -1563,348 +935,6 @@ fn parse_docker_push_digest(stdout: &str) -> Result<String> {
     })
 }
 
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct ComposeService {
-    pub name: String,
-    pub image: String,
-    /// True when `image` is the compose-generated default build tag
-    /// (`<project>-<service>:latest`) synthesized for a `build:`-only
-    /// service, rather than an explicit `image:` value.
-    pub default_image: bool,
-}
-
-/// Parsed contents of a compose config file.
-#[derive(Debug)]
-pub struct ComposeContents {
-    /// One entry per service, each resolved to an image reference: the
-    /// explicit `image:` value, or for `build:`-only services the
-    /// compose-default build tag (`<project>-<service>:latest`, flagged via
-    /// [`ComposeService::default_image`]).
-    pub services: Vec<ComposeService>,
-    /// Names of services whose image reference couldn't be resolved — no
-    /// explicit `image:` and no way to derive compose's default name. A
-    /// backstop: compose itself rejects services with neither `image` nor
-    /// `build`, and `docker-compose config` always reports a project `name`.
-    pub services_without_image: Vec<String>,
-    /// Service names that have a `build:` stanza.
-    pub build_services: HashSet<String>,
-    /// Explicitly declared network names (from the top-level `networks` key).
-    pub networks: Vec<String>,
-}
-
-/// Parse services and networks from resolved compose config YAML.
-pub fn parse_compose_config(yaml: &str) -> Result<ComposeContents> {
-    let doc: serde_yaml::Value =
-        serde_yaml::from_str(yaml).wrap_err("failed to parse docker-compose.yaml")?;
-
-    // `docker-compose config` resolves the project name (file `name:`, else
-    // the config directory's basename) and reports it at the top level. It's
-    // what compose prefixes onto default build tags, so resolution here
-    // matches what `docker compose build` produced — provided nothing
-    // overrides the project name out from under us.
-    let project_name = doc.get("name").and_then(|n| n.as_str());
-
-    let mut services = Vec::new();
-    let mut services_without_image = Vec::new();
-    let mut build_services = HashSet::new();
-    if let Some(svc_map) = doc.get("services").and_then(|s| s.as_mapping()) {
-        for (name, service) in svc_map {
-            if let Some(name_str) = name.as_str() {
-                let has_build = service.get("build").is_some();
-                if has_build {
-                    build_services.insert(name_str.to_string());
-                }
-                if let Some(image) = service.get("image").and_then(|i| i.as_str()) {
-                    services.push(ComposeService {
-                        name: name_str.to_string(),
-                        image: image.to_string(),
-                        default_image: false,
-                    });
-                } else if has_build && let Some(project) = project_name {
-                    // `docker compose build` tags a build-only service as
-                    // `<project>-<service>` (implicitly `:latest`).
-                    services.push(ComposeService {
-                        name: name_str.to_string(),
-                        image: format!("{project}-{name_str}:latest"),
-                        default_image: true,
-                    });
-                } else {
-                    services_without_image.push(name_str.to_string());
-                }
-            }
-        }
-    }
-
-    let mut networks = Vec::new();
-    if let Some(net_map) = doc.get("networks").and_then(|s| s.as_mapping()) {
-        for (name, value) in net_map {
-            if let Some(name) = name.as_str() {
-                let is_external = value
-                    .get("external")
-                    .and_then(|v| v.as_bool())
-                    .unwrap_or(false);
-                if is_external {
-                    bail!("network '{name}' is declared as external and won't work on Antithesis");
-                }
-                networks.push(name.to_string());
-            }
-        }
-    }
-    networks.sort();
-
-    Ok(ComposeContents {
-        services,
-        services_without_image,
-        build_services,
-        networks,
-    })
-}
-
-/// Ensure the referenced images are available in the local image store.
-/// snouty never pulls — what runs (validate) and what gets pushed (launch)
-/// is exactly what's in the local store.
-///
-/// The error is intentionally context-free: it explains only why local presence
-/// is required. Command-specific escape hatches (e.g. launch's `--config-image`,
-/// see [`with_config_image_escape_hatch`]) are layered on by the caller so this
-/// shared check doesn't have to know who called it.
-pub fn validate_images_are_available(
-    runtime: &dyn ContainerRuntime,
-    contents: &ComposeContents,
-) -> Result<()> {
-    ensure_services_have_images(contents)?;
-
-    let mut seen = HashSet::new();
-    let mut missing = Vec::new();
-    let mut missing_refs = Vec::new();
-
-    for service in &contents.services {
-        if !seen.insert(service.image.as_str()) {
-            continue;
-        }
-
-        if !runtime.image_exists(&service.image)? {
-            let hint = if service.default_image {
-                format!(
-                    " (compose's default build tag for service '{}' — run `docker compose build`, \
-                     or add an explicit `image:` if it was built under another name)",
-                    service.name
-                )
-            } else if contents.build_services.contains(&service.name) {
-                " (service has a `build:` stanza — build it first; snouty doesn't build)"
-                    .to_string()
-            } else {
-                String::new()
-            };
-            missing.push(format!("image: {}{hint}", service.image));
-            missing_refs.push(service.image.clone());
-        }
-    }
-
-    if missing.is_empty() {
-        return Ok(());
-    }
-
-    let mut err = eyre!("some images are not available locally");
-    for note in missing {
-        err = err.with_note(|| note);
-    }
-    err = err.with_note(|| {
-        "snouty never pulls — what it validates and launches is exactly what's in your \
-         local image store, so every referenced image must already be present there"
-    });
-
-    // A missing image is often just sitting in the *other* installed engine's
-    // store — the usual cause is `docker compose build` landing it in docker
-    // while snouty auto-selected podman (or vice versa), since the two keep
-    // separate image stores. Surface that instead of leaving the generic
-    // "build it first" note to mislead someone who already built the image.
-    // Best-effort: any probe failure counts as "not there".
-    let elsewhere = images_available_in_other_engines(runtime, &missing_refs);
-    if let Some((warnings, suggestion)) = cross_engine_guidance(runtime.name(), &elsewhere) {
-        for warning in warnings {
-            err = err.with_warning(move || warning);
-        }
-        err = err.with_suggestion(move || suggestion);
-    }
-
-    Err(err.with_suggestion(|| "pull or build the missing images, then retry"))
-}
-
-/// Probe every installed engine other than `active` for the given images.
-/// Returns, for each other engine that holds at least one, its name and the
-/// images it has. Best-effort — a probe error counts as "absent", since this
-/// only enriches an error that is already being returned.
-fn images_available_in_other_engines(
-    active: &dyn ContainerRuntime,
-    images: &[String],
-) -> Vec<(String, Vec<String>)> {
-    if images.is_empty() {
-        return Vec::new();
-    }
-    let active_name = active.name();
-    available_engines()
-        .into_iter()
-        .filter(|engine| engine.name() != active_name)
-        .filter_map(|engine| {
-            let present: Vec<String> = images
-                .iter()
-                .filter(|image| engine.image_exists(image).unwrap_or(false))
-                .cloned()
-                .collect();
-            (!present.is_empty()).then(|| (engine.name().to_string(), present))
-        })
-        .collect()
-}
-
-/// Build cross-engine guidance for images missing from the active engine
-/// (`active`) but present in another installed one. `elsewhere` pairs each
-/// other engine's name with the missing images it holds. Returns the per-image
-/// warning lines plus one suggestion pointing at the engine override, or `None`
-/// when nothing turned up elsewhere. Pure, so it is unit-tested without real
-/// engines.
-fn cross_engine_guidance(
-    active: &str,
-    elsewhere: &[(String, Vec<String>)],
-) -> Option<(Vec<String>, String)> {
-    let (source, _) = elsewhere.first()?;
-
-    let warnings = elsewhere
-        .iter()
-        .flat_map(|(engine, images)| {
-            images.iter().map(move |image| {
-                format!(
-                    "image '{image}' is in {engine}'s local image store but not {active}'s — \
-                     snouty is using {active}, and podman and docker keep separate image stores"
-                )
-            })
-        })
-        .collect();
-
-    let suggestion = format!(
-        "to use {source} instead, set SNOUTY_CONTAINER_ENGINE={source} \
-         (or add `container_engine = \"{source}\"` to a snouty settings file)"
-    );
-
-    Some((warnings, suggestion))
-}
-
-/// Layer launch's escape hatch onto a [`validate_images_are_available`] failure:
-/// a caller that already has a pre-built config image can skip local packaging
-/// entirely by launching with `--config-image <ref>`. Because `--config-image`
-/// conflicts with `--config`, the suggestion tells users to *replace* `--config`,
-/// not add the flag alongside it. Only the launch path has this alternative, so
-/// only the launch caller wraps its check with this.
-fn with_config_image_escape_hatch<T>(result: Result<T>) -> Result<T> {
-    result.with_suggestion(|| {
-        "if you already have a pre-built config image, launch with `--config-image <ref>` \
-         in place of `--config <dir>` to reuse it and skip local packaging"
-    })
-}
-
-/// Ensure the given image references all use the amd64 architecture.
-fn validate_image_architectures<R>(runtime: &R, images: &[&str]) -> Result<()>
-where
-    R: ContainerRuntime + ?Sized,
-{
-    let mut seen = HashSet::new();
-    let mut unsupported = Vec::new();
-
-    for image in images {
-        if !seen.insert(*image) {
-            continue;
-        }
-
-        let arch = runtime.image_architecture(image)?;
-        if arch != "amd64" {
-            unsupported.push(format!("image '{image}' has architecture '{arch}'"));
-        }
-    }
-
-    if unsupported.is_empty() {
-        return Ok(());
-    }
-
-    let mut err = eyre!("Antithesis requires x86-64 (amd64) container images");
-    for detail in unsupported {
-        err = err.with_note(|| detail);
-    }
-    err = err.with_suggestion(|| "use x86-64 (amd64) images, then retry");
-    Err(err)
-}
-
-/// A container reported by `compose ps`, with just the fields snouty needs.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct ComposeContainer {
-    /// Compose service name (e.g. `"app"`).
-    pub service: String,
-    /// Container ID (whatever the runtime emitted — short or full).
-    pub id: String,
-    /// Whether the container's entrypoint has exited. Antithesis can't run
-    /// test commands in stopped containers, so validate flags any service
-    /// that has test commands defined but ended up in this state. Only
-    /// `exited` and `dead` count as stopped — transient states like
-    /// `created`, `restarting`, `paused`, and missing State are treated as
-    /// not-stopped to avoid false positives on healthy setups still settling.
-    pub stopped: bool,
-}
-
-/// Determine whether a `State` field value reports a stopped container.
-/// Only `exited` and `dead` qualify — every other value (including missing
-/// `State`, transient `created`/`restarting`/`paused`, and human-readable
-/// forms like "Up 5 seconds") is treated as not-stopped.
-fn state_is_stopped(state: Option<&str>) -> bool {
-    match state {
-        Some(s) => s.eq_ignore_ascii_case("exited") || s.eq_ignore_ascii_case("dead"),
-        None => false,
-    }
-}
-
-/// Parse the JSON output of `docker-compose ps --format json`.
-///
-/// Handles both NDJSON (one object per line) and JSON array formats. The
-/// schema is Docker Compose v2: `{"Service": "...", "ID": "...", "State": "running"}`.
-fn parse_compose_ps(stdout: &str) -> Result<Vec<ComposeContainer>> {
-    let stdout = stdout.trim();
-    if stdout.is_empty() {
-        return Ok(Vec::new());
-    }
-
-    let entries: Vec<serde_json::Value> = if stdout.starts_with('[') {
-        serde_json::from_str(stdout).wrap_err("failed to parse compose ps JSON array")?
-    } else {
-        stdout
-            .lines()
-            .filter(|l| !l.trim().is_empty())
-            .map(serde_json::from_str)
-            .collect::<std::result::Result<Vec<_>, _>>()
-            .wrap_err("failed to parse compose ps NDJSON")?
-    };
-
-    entries
-        .iter()
-        .map(|v| {
-            let id = v
-                .get("ID")
-                .and_then(|v| v.as_str())
-                .ok_or_else(|| eyre!("missing container ID in compose ps output"))?;
-
-            let service = v
-                .get("Service")
-                .and_then(|v| v.as_str())
-                .ok_or_else(|| eyre!("missing service name in compose ps output"))?;
-
-            let state = v.get("State").and_then(|v| v.as_str());
-
-            Ok(ComposeContainer {
-                service: service.to_string(),
-                id: id.to_string(),
-                stopped: state_is_stopped(state),
-            })
-        })
-        .collect()
-}
-
 fn image_exists_from_inspect_output(
     runtime: &str,
     image_ref: &str,
@@ -1935,13 +965,10 @@ fn image_inspect_reports_missing_image(stderr: &str) -> bool {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::testutils::{
-        OCIRegistry, has_compose, require_runtimes, require_runtimes_with_compose, skip_or_fail,
-    };
+    use crate::testutils::{OCIRegistry, require_runtimes};
     use hegel::generators::{self, Generator};
-    use std::collections::BTreeMap;
+
     use std::os::unix::process::ExitStatusExt;
-    use std::path::PathBuf;
 
     /// A text generator whose output never contains `@` — for digest values,
     /// which never carry an `@` in practice (`sha256:...`), so the
@@ -2052,103 +1079,6 @@ mod tests {
         }
     }
 
-    fn cc(service: &str, id: &str, stopped: bool) -> ComposeContainer {
-        ComposeContainer {
-            service: service.to_string(),
-            id: id.to_string(),
-            stopped,
-        }
-    }
-
-    #[test]
-    fn parse_compose_ps_ndjson() {
-        let stdout = "{\"ID\":\"abc123\",\"Service\":\"app\",\"State\":\"running\"}\n\
-                      {\"ID\":\"def456\",\"Service\":\"sidecar\",\"State\":\"exited\"}\n";
-        let result = parse_compose_ps(stdout).unwrap();
-        assert_eq!(
-            result,
-            vec![cc("app", "abc123", false), cc("sidecar", "def456", true)]
-        );
-    }
-
-    #[test]
-    fn parse_compose_ps_json_array() {
-        let stdout = r#"[
-            {"ID":"abc123","Service":"app","State":"running"},
-            {"ID":"def456","Service":"sidecar","State":"running"}
-        ]"#;
-        let result = parse_compose_ps(stdout).unwrap();
-        assert_eq!(
-            result,
-            vec![cc("app", "abc123", false), cc("sidecar", "def456", false)]
-        );
-    }
-
-    #[test]
-    fn parse_compose_ps_empty() {
-        let result = parse_compose_ps("").unwrap();
-        assert!(result.is_empty());
-    }
-
-    #[test]
-    fn parse_compose_ps_missing_state_is_not_stopped() {
-        // A container with no State field — typical during early startup —
-        // must NOT be classified as stopped, or the stranded-test-commands
-        // diagnostic fires on healthy containers.
-        let stdout = r#"[{"ID":"abc","Service":"app"}]"#;
-        let result = parse_compose_ps(stdout).unwrap();
-        assert_eq!(result, vec![cc("app", "abc", false)]);
-    }
-
-    #[test]
-    fn parse_compose_ps_transient_states_are_not_stopped() {
-        // created / restarting / paused are not "stopped" — Antithesis may
-        // still see them recover. Only `exited` and `dead` count as stopped.
-        let stdout = r#"[
-            {"ID":"a","Service":"svc","State":"created"},
-            {"ID":"b","Service":"svc","State":"restarting"},
-            {"ID":"c","Service":"svc","State":"paused"},
-            {"ID":"d","Service":"svc","State":"Up 5 seconds"},
-            {"ID":"e","Service":"svc","State":"dead"},
-            {"ID":"f","Service":"svc","State":"EXITED"}
-        ]"#;
-        let result = parse_compose_ps(stdout).unwrap();
-        let stopped: Vec<(&str, bool)> =
-            result.iter().map(|c| (c.id.as_str(), c.stopped)).collect();
-        assert_eq!(
-            stopped,
-            vec![
-                ("a", false),
-                ("b", false),
-                ("c", false),
-                ("d", false),
-                ("e", true),
-                ("f", true),
-            ]
-        );
-    }
-
-    #[test]
-    fn parse_compose_ps_returns_one_entry_per_replica() {
-        // Scaled services (`replicas: N`) emit one entry per container, all
-        // sharing the same Service value but with distinct IDs. Validate
-        // keys per-container work by container.id rather than service.
-        let stdout = r#"[
-            {"ID":"a1","Service":"worker","State":"running"},
-            {"ID":"a2","Service":"worker","State":"running"},
-            {"ID":"a3","Service":"worker","State":"exited"}
-        ]"#;
-        let result = parse_compose_ps(stdout).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                cc("worker", "a1", false),
-                cc("worker", "a2", false),
-                cc("worker", "a3", true),
-            ]
-        );
-    }
-
     #[test]
     fn stderr_indicates_missing_source_matches_docker_and_podman() {
         assert!(stderr_indicates_missing_source(
@@ -2195,290 +1125,6 @@ mod tests {
         assert!(
             image_ref.starts_with("registry.example.com/repo/snouty-config:"),
             "got: {image_ref}"
-        );
-    }
-
-    #[test]
-    fn parse_compose_config_basic() {
-        let yaml = "\
-services:
-  app:
-    image: us-central1-docker.pkg.dev/proj/repo/app:v1
-  sidecar:
-    image: us-central1-docker.pkg.dev/proj/repo/sidecar:latest
-  builder:
-    build:
-      context: ./builder
-";
-        let contents = parse_compose_config(yaml).unwrap();
-        assert_eq!(
-            contents.services,
-            vec![
-                ComposeService {
-                    name: "app".to_string(),
-                    image: "us-central1-docker.pkg.dev/proj/repo/app:v1".to_string(),
-                    default_image: false,
-                },
-                ComposeService {
-                    name: "sidecar".to_string(),
-                    image: "us-central1-docker.pkg.dev/proj/repo/sidecar:latest".to_string(),
-                    default_image: false,
-                },
-            ]
-        );
-        assert_eq!(
-            contents.build_services,
-            HashSet::from(["builder".to_string()])
-        );
-        // No top-level `name:` → the builder service can't be resolved to
-        // compose's default build tag.
-        assert_eq!(contents.services_without_image, vec!["builder".to_string()]);
-        assert!(contents.networks.is_empty());
-    }
-
-    #[test]
-    fn parse_compose_config_synthesizes_default_build_tags() {
-        // `docker-compose config` output always carries the resolved project
-        // name; build-only services resolve to `<project>-<service>:latest`,
-        // exactly the tag `docker compose build` produces.
-        let yaml = "\
-name: myproj
-services:
-  app:
-    image: myapp:latest
-  builder:
-    build:
-      context: ./builder
-";
-        let contents = parse_compose_config(yaml).unwrap();
-        assert_eq!(
-            contents.services,
-            vec![
-                ComposeService {
-                    name: "app".to_string(),
-                    image: "myapp:latest".to_string(),
-                    default_image: false,
-                },
-                ComposeService {
-                    name: "builder".to_string(),
-                    image: "myproj-builder:latest".to_string(),
-                    default_image: true,
-                },
-            ]
-        );
-        assert!(contents.services_without_image.is_empty());
-    }
-
-    #[test]
-    fn parse_compose_config_no_services() {
-        let yaml = "version: '3'\n";
-        let contents = parse_compose_config(yaml).unwrap();
-        assert!(contents.services.is_empty());
-        assert!(contents.build_services.is_empty());
-    }
-
-    #[test]
-    fn parse_compose_config_with_networks() {
-        let yaml = "\
-services:
-  app:
-    image: myapp:latest
-networks:
-  backend: {}
-  frontend:
-    driver: bridge
-";
-        let contents = parse_compose_config(yaml).unwrap();
-        assert_eq!(contents.services.len(), 1);
-        assert!(contents.build_services.is_empty());
-        assert_eq!(contents.networks, vec!["backend", "frontend"]);
-    }
-
-    #[test]
-    fn parse_compose_config_rejects_external_network() {
-        let yaml = "\
-services:
-  app:
-    image: myapp:latest
-networks:
-  shared_net:
-    external: true
-";
-        let err = parse_compose_config(yaml).unwrap_err();
-        assert!(
-            err.to_string().contains("external"),
-            "expected error about external network, got: {err}"
-        );
-    }
-
-    #[test]
-    fn compose_config_resolves_env() {
-        let runtimes = require_runtimes_with_compose();
-        if runtimes.is_empty() {
-            return;
-        }
-
-        for rt in &runtimes {
-            eprintln!("testing with runtime: {}", rt.name());
-            let dir = tempfile::tempdir().unwrap();
-            std::fs::write(
-                dir.path().join(".env"),
-                "REPOSITORY=us-central1-docker.pkg.dev/proj/repo\nIMAGES_TAG=v2\n",
-            )
-            .unwrap();
-            std::fs::write(
-                dir.path().join("docker-compose.yaml"),
-                "\
-services:
-  app:
-    image: ${REPOSITORY}/app:${IMAGES_TAG}
-  sidecar:
-    image: docker.io/library/nginx:latest
-",
-            )
-            .unwrap();
-
-            let compose = DockerCompose::resolve(rt.as_ref()).unwrap();
-            let config = match crate::config::detect_config(dir.path()).unwrap() {
-                crate::config::Config::Compose(c) => c,
-                other => panic!("expected Compose, got {other:?}"),
-            };
-            let contents = compose.contents(&config, None).unwrap();
-            let images: Vec<&str> = contents
-                .services
-                .iter()
-                .map(|service| service.image.as_str())
-                .collect();
-            assert_eq!(
-                images,
-                vec![
-                    "us-central1-docker.pkg.dev/proj/repo/app:v2",
-                    "docker.io/library/nginx:latest",
-                ],
-                "failed for runtime: {}",
-                rt.name()
-            );
-        }
-    }
-
-    #[test]
-    fn compose_contents_apply_overlays() {
-        let runtimes = require_runtimes_with_compose();
-        if runtimes.is_empty() {
-            return;
-        }
-
-        for rt in &runtimes {
-            eprintln!("testing with runtime: {}", rt.name());
-            let dir = tempfile::tempdir().unwrap();
-            std::fs::write(
-                dir.path().join("docker-compose.yaml"),
-                "\
-services:
-  app:
-    image: base:latest
-",
-            )
-            .unwrap();
-            let overlay = dir.path().join("override.yaml");
-            std::fs::write(
-                &overlay,
-                "\
-services:
-  app:
-    image: overlay:latest
-",
-            )
-            .unwrap();
-
-            let compose = DockerCompose::resolve(rt.as_ref()).unwrap();
-            let config = match crate::config::detect_config(dir.path()).unwrap() {
-                crate::config::Config::Compose(c) => c,
-                other => panic!("expected Compose, got {other:?}"),
-            };
-            let yaml = compose.config_yaml(&config, Some(&overlay), &[]).unwrap();
-            let contents = compose.contents(&config, Some(&overlay)).unwrap();
-
-            assert!(
-                yaml.contains("overlay:latest"),
-                "expected overlay image in resolved yaml for runtime {}: {yaml}",
-                rt.name()
-            );
-            assert_eq!(contents.services.len(), 1);
-            assert_eq!(contents.services[0].image, "overlay:latest");
-        }
-    }
-
-    #[test]
-    fn rewrite_compose_images_pins_every_service() {
-        // Input mirrors `docker-compose config --no-interpolate` output:
-        // machine-generated YAML where every service must end up pinned.
-        // Non-image fields (build, volumes, environment) are preserved.
-        let yaml = "\
-services:
-  app:
-    build:
-      context: .
-    image: ${REPO}/app:${TAG}
-    volumes:
-      - ./data:/data
-  nginx:
-    image: docker.io/library/nginx:latest
-    environment:
-      FOO: bar
-";
-        let pinned = BTreeMap::from([
-            (
-                "app".to_string(),
-                "reg.example.com/app:v1@sha256:aaa".to_string(),
-            ),
-            (
-                "nginx".to_string(),
-                "docker.io/library/nginx:latest@sha256:bbb".to_string(),
-            ),
-        ]);
-
-        let out = rewrite_compose_images(yaml, &pinned).unwrap();
-        let doc: serde_yaml::Value = serde_yaml::from_str(&out).unwrap();
-        let services = doc.get("services").unwrap();
-
-        let image = |svc: &str| {
-            services
-                .get(svc)
-                .and_then(|s| s.get("image"))
-                .and_then(|i| i.as_str())
-                .unwrap()
-                .to_string()
-        };
-        assert_eq!(image("app"), "reg.example.com/app:v1@sha256:aaa");
-        assert_eq!(image("nginx"), "docker.io/library/nginx:latest@sha256:bbb");
-        // Surrounding structure is preserved.
-        assert!(services.get("app").unwrap().get("build").is_some());
-        assert!(services.get("app").unwrap().get("volumes").is_some());
-        assert!(services.get("nginx").unwrap().get("environment").is_some());
-    }
-
-    #[test]
-    fn rewrite_compose_images_rejects_unpinned_service() {
-        // A service the pinning pass lost track of must fail loudly instead of
-        // shipping an unpinned image reference to the platform.
-        let yaml = "\
-services:
-  app:
-    image: app:latest
-  forgotten:
-    image: forgotten:latest
-";
-        let pinned = BTreeMap::from([(
-            "app".to_string(),
-            "reg.example.com/app:latest@sha256:aaa".to_string(),
-        )]);
-
-        let err = rewrite_compose_images(yaml, &pinned).unwrap_err();
-        let msg = format!("{err:?}");
-        assert!(
-            msg.contains("'forgotten'") && msg.contains("bug in snouty"),
-            "expected the unpinned service to be flagged as a bug, got: {msg}"
         );
     }
 
@@ -2548,573 +1194,6 @@ services:
         assert_eq!(
             classify_manifest_json(b"not json"),
             RemoteManifest::NotFound
-        );
-    }
-
-    /// Run pin_images over `yaml` with a [`FakeRuntime`] (real docker-compose
-    /// binary for config resolution, fake image/registry operations).
-    fn pin_with_fake(rt: &FakeRuntime, yaml: &str, registry: &str) -> Result<String> {
-        let dir = tempfile::tempdir().unwrap();
-        std::fs::write(dir.path().join("docker-compose.yaml"), yaml).unwrap();
-        let config = match crate::config::detect_config(dir.path()).unwrap() {
-            crate::config::Config::Compose(c) => c,
-            other => panic!("expected Compose, got {other:?}"),
-        };
-        let compose = DockerCompose::resolve(rt).unwrap();
-        compose.pin_images(rt, &config, registry)
-    }
-
-    #[test]
-    fn pin_images_skips_push_when_registry_serves_digest() {
-        if !has_compose() {
-            // Loud in CI (skip_or_fail panics there) so a runner missing
-            // docker-compose can't silently drop this coverage.
-            skip_or_fail("docker-compose (Docker Compose v2) is not available");
-            return;
-        }
-        // A pulled third-party image: the registry confirms the local digest
-        // (a multi-arch list with amd64), so it's pinned there without a push.
-        let rt = FakeRuntime {
-            available_images: BTreeMap::from([("redis:7".to_string(), true)]),
-            repo_digests: BTreeMap::from([(
-                "redis:7".to_string(),
-                vec![
-                    // The per-arch entry can't be verified (podman can't
-                    // inspect single manifests) — the list entry can.
-                    "docker.io/library/redis@sha256:child".to_string(),
-                    "docker.io/library/redis@sha256:list".to_string(),
-                ],
-            )]),
-            remote_manifests: BTreeMap::from([(
-                "docker.io/library/redis@sha256:list".to_string(),
-                RemoteManifest::List { has_amd64: true },
-            )]),
-            ..Default::default()
-        };
-        let out = pin_with_fake(
-            &rt,
-            "services:\n  app:\n    image: redis:7\n",
-            "reg.example.com",
-        )
-        .unwrap();
-        assert!(
-            out.contains("docker.io/library/redis:7@sha256:list"),
-            "expected the verified list digest pin, got: {out}"
-        );
-        assert!(
-            rt.pushed.lock().unwrap().is_empty(),
-            "nothing should be pushed"
-        );
-    }
-
-    #[test]
-    fn pin_images_pushes_when_no_registry_serves_digest() {
-        if !has_compose() {
-            // Loud in CI (skip_or_fail panics there) so a runner missing
-            // docker-compose can't silently drop this coverage.
-            skip_or_fail("docker-compose (Docker Compose v2) is not available");
-            return;
-        }
-        // The local store fabricates digest entries for registry-qualified
-        // tags that were never pushed; the registry round trip rejects them
-        // (NotFound) and the image is pushed to our registry instead.
-        let rt = FakeRuntime {
-            available_images: BTreeMap::from([("ghcr.io/org/app:v1".to_string(), true)]),
-            architectures: BTreeMap::from([(
-                "reg.example.com/ghcr.io/org/app:v1".to_string(),
-                "amd64".to_string(),
-            )]),
-            repo_digests: BTreeMap::from([(
-                "ghcr.io/org/app:v1".to_string(),
-                vec!["ghcr.io/org/app@sha256:fabricated".to_string()],
-            )]),
-            ..Default::default()
-        };
-        let out = pin_with_fake(
-            &rt,
-            "services:\n  app:\n    image: ghcr.io/org/app:v1\n",
-            "reg.example.com",
-        )
-        .unwrap();
-        assert!(
-            out.contains("reg.example.com/ghcr.io/org/app:v1@sha256:fakepushdigest"),
-            "expected push-pinned reference, got: {out}"
-        );
-        assert_eq!(
-            *rt.pushed.lock().unwrap(),
-            vec!["reg.example.com/ghcr.io/org/app:v1".to_string()]
-        );
-    }
-
-    #[test]
-    fn pin_images_skips_push_for_previously_pushed_image() {
-        if !has_compose() {
-            // Loud in CI (skip_or_fail panics there) so a runner missing
-            // docker-compose can't silently drop this coverage.
-            skip_or_fail("docker-compose (Docker Compose v2) is not available");
-            return;
-        }
-        // A bare local image pushed by an earlier launch: the
-        // registry-prefixed candidate verifies, so no re-push. The manifest
-        // is single-platform, so the local architecture must be amd64.
-        let rt = FakeRuntime {
-            available_images: BTreeMap::from([("myapp:latest".to_string(), true)]),
-            architectures: BTreeMap::from([("myapp:latest".to_string(), "amd64".to_string())]),
-            repo_digests: BTreeMap::from([(
-                "myapp:latest".to_string(),
-                vec!["reg.example.com/myapp@sha256:pushedearlier".to_string()],
-            )]),
-            remote_manifests: BTreeMap::from([(
-                "reg.example.com/myapp@sha256:pushedearlier".to_string(),
-                RemoteManifest::Single,
-            )]),
-            ..Default::default()
-        };
-        let out = pin_with_fake(
-            &rt,
-            "services:\n  app:\n    image: myapp:latest\n",
-            "reg.example.com",
-        )
-        .unwrap();
-        assert!(
-            out.contains("reg.example.com/myapp:latest@sha256:pushedearlier"),
-            "expected pin to the previously pushed digest, got: {out}"
-        );
-        assert!(
-            rt.pushed.lock().unwrap().is_empty(),
-            "nothing should be pushed"
-        );
-    }
-
-    #[test]
-    fn pin_images_rejects_arm_only_images() {
-        if !has_compose() {
-            // Loud in CI (skip_or_fail panics there) so a runner missing
-            // docker-compose can't silently drop this coverage.
-            skip_or_fail("docker-compose (Docker Compose v2) is not available");
-            return;
-        }
-        // The registry serves the digest, but only as arm (a list without an
-        // amd64 entry). The pin is refused and the push path's local arch
-        // check produces the amd64 guidance before anything is pushed.
-        let rt = FakeRuntime {
-            available_images: BTreeMap::from([("armthing:latest".to_string(), true)]),
-            architectures: BTreeMap::from([
-                ("armthing:latest".to_string(), "arm64".to_string()),
-                (
-                    "reg.example.com/armthing:latest".to_string(),
-                    "arm64".to_string(),
-                ),
-            ]),
-            repo_digests: BTreeMap::from([(
-                "armthing:latest".to_string(),
-                vec!["docker.io/library/armthing@sha256:armlist".to_string()],
-            )]),
-            remote_manifests: BTreeMap::from([(
-                "docker.io/library/armthing@sha256:armlist".to_string(),
-                RemoteManifest::List { has_amd64: false },
-            )]),
-            ..Default::default()
-        };
-        let err = pin_with_fake(
-            &rt,
-            "services:\n  app:\n    image: armthing:latest\n",
-            "reg.example.com",
-        )
-        .unwrap_err();
-        let debug = format!("{err:?}");
-        assert!(
-            debug.contains("amd64"),
-            "expected amd64 guidance, got: {debug}"
-        );
-        assert!(
-            rt.pushed.lock().unwrap().is_empty(),
-            "nothing should be pushed"
-        );
-    }
-
-    #[test]
-    fn pin_images_pushes_every_local_image() {
-        let runtimes = require_runtimes_with_compose();
-        if runtimes.is_empty() {
-            return;
-        }
-
-        for rt in &runtimes {
-            eprintln!("testing with runtime: {}", rt.name());
-            let registry = match OCIRegistry::start(rt.as_ref()) {
-                Some(r) => r,
-                None => continue,
-            };
-            let addr = registry.host_port();
-            let compose = DockerCompose::resolve(rt.as_ref())
-                .unwrap_or_else(|e| panic!("{}: DockerCompose::resolve: {e:?}", rt.name()));
-
-            // Build a purely-local image (present locally, in no registry).
-            let img_dir = tempfile::tempdir().unwrap();
-            std::fs::write(
-                img_dir.path().join("Dockerfile"),
-                "FROM scratch\nCOPY . /\n",
-            )
-            .unwrap();
-            std::fs::write(img_dir.path().join("file"), "x").unwrap();
-            let local = "snouty-pin-test:latest";
-            rt.build_image(img_dir.path(), local, None, Some("linux/amd64"))
-                .unwrap_or_else(|e| panic!("{}: build: {e:?}", rt.name()));
-
-            // Resolve `app`'s pinned image after running pin_images over `yaml`.
-            let pinned_app = |yaml: &str| -> Result<String> {
-                let dir = tempfile::tempdir().unwrap();
-                std::fs::write(dir.path().join("docker-compose.yaml"), yaml).unwrap();
-                let config = match crate::config::detect_config(dir.path()).unwrap() {
-                    crate::config::Config::Compose(c) => c,
-                    other => panic!("expected Compose, got {other:?}"),
-                };
-                let out = compose.pin_images(rt.as_ref(), &config, &addr)?;
-                Ok(serde_yaml::from_str::<serde_yaml::Value>(&out)
-                    .unwrap()
-                    .get("services")
-                    .and_then(|s| s.get("app"))
-                    .and_then(|s| s.get("image"))
-                    .and_then(|i| i.as_str())
-                    .unwrap()
-                    .to_string())
-            };
-            let pushed_prefix = format!("{addr}/snouty-pin-test:latest@sha256:");
-
-            // Case 1 — build stanza: the local build is pushed and pinned.
-            let built = pinned_app(&format!(
-                "services:\n  app:\n    build: .\n    image: {local}\n"
-            ))
-            .unwrap_or_else(|e| panic!("{}: build case: {e:?}", rt.name()));
-            assert!(
-                built.starts_with(&pushed_prefix),
-                "{}: build image should be pushed, got: {built}",
-                rt.name()
-            );
-
-            // Case 2 — local without a build stanza (prebuilt/loaded out of
-            // band): local availability is enough; pushed and pinned the same.
-            let local_only = pinned_app(&format!("services:\n  app:\n    image: {local}\n"))
-                .unwrap_or_else(|e| panic!("{}: local-only case: {e:?}", rt.name()));
-            assert!(
-                local_only.starts_with(&pushed_prefix),
-                "{}: local-only image should be pushed, got: {local_only}",
-                rt.name()
-            );
-
-            // Case 3 — not present locally: hard error before anything is
-            // pushed. snouty never pulls, even for registry-qualified refs.
-            let err = pinned_app("services:\n  app:\n    image: snouty-bare-local-xyz:latest\n")
-                .expect_err(&format!("{}: expected pin_images to fail", rt.name()));
-            let debug = format!("{err:?}");
-            assert!(
-                debug.contains("snouty-bare-local-xyz:latest")
-                    && debug.contains("not available locally"),
-                "{}: error should name the missing image, got: {debug}",
-                rt.name()
-            );
-
-            // Case 4 — `build:`-only service with no `image:` resolves to
-            // compose's default build tag (`<project>-<service>:latest`);
-            // when that tag was never built, the error names it with
-            // guidance instead of silently launching an image the platform
-            // can't pull.
-            let err = pinned_app("services:\n  app:\n    build: .\n").expect_err(&format!(
-                "{}: expected unbuilt default-tag service to fail",
-                rt.name()
-            ));
-            let debug = format!("{err:?}");
-            assert!(
-                debug.contains("-app:latest") && debug.contains("default build tag"),
-                "{}: error should name the default build tag, got: {debug}",
-                rt.name()
-            );
-
-            let _ = Command::new(rt.name())
-                .args(["rmi", local, &format!("{addr}/{local}")])
-                .output();
-        }
-    }
-
-    #[derive(Clone, Default)]
-    struct FakeRuntime {
-        available_images: BTreeMap<String, bool>,
-        architectures: BTreeMap<String, String>,
-        repo_digests: BTreeMap<String, Vec<String>>,
-        remote_manifests: BTreeMap<String, RemoteManifest>,
-        pushed: std::sync::Arc<std::sync::Mutex<Vec<String>>>,
-    }
-
-    impl ContainerRuntime for FakeRuntime {
-        fn name(&self) -> &str {
-            "fake"
-        }
-
-        fn clone_box(&self) -> Box<dyn ContainerRuntime> {
-            Box::new(self.clone())
-        }
-
-        fn image_push(&self, image_ref: &str) -> Result<String> {
-            self.pushed.lock().unwrap().push(image_ref.to_string());
-            Ok(pinned_image_ref(image_ref, "sha256:fakepushdigest"))
-        }
-
-        fn image_exists(&self, image_ref: &str) -> Result<bool> {
-            Ok(*self.available_images.get(image_ref).unwrap_or(&false))
-        }
-
-        fn image_architecture(&self, image_ref: &str) -> Result<String> {
-            self.architectures
-                .get(image_ref)
-                .cloned()
-                .ok_or_else(|| eyre!("missing fake architecture for {image_ref}"))
-        }
-
-        fn image_repo_digests(&self, image_ref: &str) -> Result<Vec<String>> {
-            Ok(self
-                .repo_digests
-                .get(image_ref)
-                .cloned()
-                .unwrap_or_default())
-        }
-
-        fn image_tag(&self, _src: &str, _dst: &str) -> Result<()> {
-            Ok(())
-        }
-
-        fn remote_manifest(&self, image_ref: &str) -> RemoteManifest {
-            self.remote_manifests
-                .get(image_ref)
-                .cloned()
-                .unwrap_or(RemoteManifest::NotFound)
-        }
-    }
-
-    #[test]
-    fn validate_image_architectures_accepts_amd64_images() {
-        let runtime = FakeRuntime {
-            available_images: BTreeMap::new(),
-            architectures: BTreeMap::from([
-                ("app:latest".to_string(), "amd64".to_string()),
-                ("sidecar:latest".to_string(), "amd64".to_string()),
-            ]),
-            ..Default::default()
-        };
-
-        validate_image_architectures(&runtime, &["app:latest", "sidecar:latest"]).unwrap();
-    }
-
-    #[test]
-    fn validate_image_architectures_rejects_non_amd64_images() {
-        let runtime = FakeRuntime {
-            available_images: BTreeMap::new(),
-            architectures: BTreeMap::from([
-                ("app:latest".to_string(), "arm64".to_string()),
-                ("sidecar:latest".to_string(), "amd64".to_string()),
-            ]),
-            ..Default::default()
-        };
-
-        let err =
-            validate_image_architectures(&runtime, &["app:latest", "sidecar:latest"]).unwrap_err();
-
-        let msg = err.to_string();
-        let debug = format!("{err:?}");
-        assert!(
-            msg.contains("x86-64 (amd64)"),
-            "expected architecture guidance, got: {msg}"
-        );
-        assert!(
-            debug.contains("image 'app:latest' has architecture 'arm64'"),
-            "expected offending image details, got: {debug}"
-        );
-    }
-
-    /// Build a [`ComposeContents`] from `(service, image)` pairs and the names
-    /// of services that have a `build:` stanza.
-    fn contents_of(services: &[(&str, &str)], build_services: &[&str]) -> ComposeContents {
-        ComposeContents {
-            services: services
-                .iter()
-                .map(|(name, image)| ComposeService {
-                    name: name.to_string(),
-                    image: image.to_string(),
-                    default_image: false,
-                })
-                .collect(),
-            services_without_image: Vec::new(),
-            build_services: build_services.iter().map(|s| s.to_string()).collect(),
-            networks: Vec::new(),
-        }
-    }
-
-    #[test]
-    fn validate_images_are_available_accepts_local_images() {
-        let runtime = FakeRuntime {
-            available_images: BTreeMap::from([
-                ("app:latest".to_string(), true),
-                ("sidecar:latest".to_string(), true),
-            ]),
-            architectures: BTreeMap::new(),
-            ..Default::default()
-        };
-
-        validate_images_are_available(
-            &runtime,
-            &contents_of(&[("app", "app:latest"), ("sidecar", "sidecar:latest")], &[]),
-        )
-        .unwrap();
-    }
-
-    #[test]
-    fn validate_images_are_available_reports_all_missing_images() {
-        let runtime = FakeRuntime {
-            available_images: BTreeMap::from([
-                ("present:latest".to_string(), true),
-                ("missing-a:latest".to_string(), false),
-                ("missing-b:latest".to_string(), false),
-            ]),
-            architectures: BTreeMap::new(),
-            ..Default::default()
-        };
-
-        let err = validate_images_are_available(
-            &runtime,
-            &contents_of(
-                &[
-                    ("present", "present:latest"),
-                    ("app", "missing-a:latest"),
-                    ("sidecar", "missing-b:latest"),
-                ],
-                &["sidecar"],
-            ),
-        )
-        .unwrap_err();
-
-        let msg = err.to_string();
-        let debug = format!("{err:?}");
-        assert!(
-            msg.contains("some images are not available locally"),
-            "expected missing-image guidance, got: {msg}"
-        );
-        assert!(
-            debug.contains("image: missing-a:latest"),
-            "expected first missing image details, got: {debug}"
-        );
-        assert!(
-            debug.contains("image: missing-b:latest (service has a `build:` stanza"),
-            "expected build-stanza hint on the second missing image, got: {debug}"
-        );
-        assert!(
-            debug.contains("snouty never pulls"),
-            "expected the why-it's-required-locally note, got: {debug}"
-        );
-        // The shared check is context-free: the launch-only `--config-image`
-        // escape hatch is layered on by the caller, not emitted here.
-        assert!(
-            !debug.contains("--config-image"),
-            "shared check should stay context-free, got: {debug}"
-        );
-    }
-
-    #[test]
-    fn with_config_image_escape_hatch_tells_users_to_replace_config() {
-        let runtime = FakeRuntime {
-            available_images: BTreeMap::from([("missing:latest".to_string(), false)]),
-            architectures: BTreeMap::new(),
-            ..Default::default()
-        };
-
-        let err = with_config_image_escape_hatch(validate_images_are_available(
-            &runtime,
-            &contents_of(&[("app", "missing:latest")], &[]),
-        ))
-        .unwrap_err();
-
-        let debug = format!("{err:?}");
-        assert!(
-            debug.contains("--config-image <ref>"),
-            "expected the config-image escape hatch, got: {debug}"
-        );
-        // --config-image conflicts with --config, so the hint must say to replace
-        // it, not add it alongside (which clap would reject).
-        assert!(
-            debug.contains("in place of `--config <dir>`"),
-            "expected the hint to replace --config, not add it, got: {debug}"
-        );
-    }
-
-    #[test]
-    fn validate_images_are_available_rejects_imageless_services() {
-        let runtime = FakeRuntime {
-            available_images: BTreeMap::new(),
-            architectures: BTreeMap::new(),
-            ..Default::default()
-        };
-
-        let mut contents = contents_of(&[], &["app"]);
-        contents.services_without_image = vec!["app".to_string()];
-
-        let err = validate_images_are_available(&runtime, &contents).unwrap_err();
-        let debug = format!("{err:?}");
-        assert!(
-            debug.contains("service 'app' has no `image:` field"),
-            "expected imageless-service guidance, got: {debug}"
-        );
-        // The imageless check bails before the missing-local-image guidance, so
-        // its unrelated "pull or build" suggestion must not leak onto this error.
-        assert!(
-            !debug.contains("pull or build the missing images"),
-            "imageless error should not carry missing-image guidance, got: {debug}"
-        );
-    }
-
-    #[test]
-    fn cross_engine_guidance_is_none_when_nothing_found_elsewhere() {
-        assert!(cross_engine_guidance("podman", &[]).is_none());
-    }
-
-    #[test]
-    fn cross_engine_guidance_names_the_engine_and_override() {
-        let elsewhere = vec![(
-            "docker".to_string(),
-            vec!["local-benchmark-driver:local".to_string()],
-        )];
-        let (warnings, suggestion) = cross_engine_guidance("podman", &elsewhere).unwrap();
-
-        assert_eq!(warnings.len(), 1);
-        assert!(
-            warnings[0].contains("local-benchmark-driver:local")
-                && warnings[0].contains("docker's local image store")
-                && warnings[0].contains("snouty is using podman"),
-            "unexpected warning: {}",
-            warnings[0]
-        );
-        assert!(
-            suggestion.contains("SNOUTY_CONTAINER_ENGINE=docker")
-                && suggestion.contains("container_engine = \"docker\""),
-            "expected the engine override and setting, got: {suggestion}"
-        );
-        // We point at the override, not at copying images between stores.
-        assert!(
-            !suggestion.contains("save") && !suggestion.contains("load"),
-            "suggestion should not include a copy command, got: {suggestion}"
-        );
-    }
-
-    #[test]
-    fn cross_engine_guidance_warns_once_per_image() {
-        let elsewhere = vec![(
-            "docker".to_string(),
-            vec!["a:latest".to_string(), "b:latest".to_string()],
-        )];
-        let (warnings, suggestion) = cross_engine_guidance("podman", &elsewhere).unwrap();
-
-        assert_eq!(warnings.len(), 2);
-        assert!(
-            suggestion.contains("SNOUTY_CONTAINER_ENGINE=docker"),
-            "expected the engine override, got: {suggestion}"
         );
     }
 
@@ -3294,119 +1373,6 @@ tag2: digest: sha256:bbb222 size: 200
         assert!(
             debug.contains("Cannot connect to the Docker daemon"),
             "expected daemon error details, got: {debug}"
-        );
-    }
-
-    #[test]
-    fn parse_compose_config_build_with_image() {
-        let yaml = "\
-services:
-  app:
-    build: .
-    image: myapp:latest
-  sidecar:
-    image: docker.io/library/nginx:latest
-";
-        let contents = parse_compose_config(yaml).unwrap();
-        assert_eq!(
-            contents.services,
-            vec![
-                ComposeService {
-                    name: "app".to_string(),
-                    image: "myapp:latest".to_string(),
-                    default_image: false,
-                },
-                ComposeService {
-                    name: "sidecar".to_string(),
-                    image: "docker.io/library/nginx:latest".to_string(),
-                    default_image: false,
-                },
-            ]
-        );
-        assert_eq!(contents.build_services, HashSet::from(["app".to_string()]));
-    }
-
-    #[test]
-    fn is_compose_v2_version_accepts_docker_compose() {
-        assert!(is_compose_v2_version("Docker Compose version v2.24.5"));
-        assert!(is_compose_v2_version("docker compose version 5.1.4"));
-    }
-
-    #[test]
-    fn is_compose_v2_version_rejects_others() {
-        assert!(!is_compose_v2_version("podman-compose version 1.0.6"));
-        // Compose v1's hyphenated banner must not be mistaken for v2.
-        assert!(!is_compose_v2_version("docker-compose version 1.29.2"));
-        assert!(!is_compose_v2_version(""));
-    }
-
-    #[test]
-    fn compose_cli_display_standalone_and_plugin() {
-        let standalone = ComposeCli::Standalone(PathBuf::from("/usr/local/bin/docker-compose"));
-        assert_eq!(standalone.display(), "docker-compose");
-
-        let plugin = ComposeCli::Plugin(PathBuf::from("/usr/bin/docker"));
-        assert_eq!(plugin.display(), "docker compose");
-    }
-
-    #[test]
-    fn config_json_hermetic_env_scrubs_process_variables() {
-        if !has_compose() {
-            skip_or_fail("docker-compose (Docker Compose v2) is not available");
-            return;
-        }
-
-        let dir = tempfile::tempdir().unwrap();
-        std::fs::write(
-            dir.path().join("docker-compose.yaml"),
-            "\
-services:
-  app:
-    image: alpine
-    environment:
-      HOME_VALUE: \"${HOME}\"
-      PATH_VALUE: \"${PATH}\"
-      DOCKER_HOST_VALUE: \"${DOCKER_HOST}\"
-",
-        )
-        .unwrap();
-        let config = match crate::config::detect_config(dir.path()).unwrap() {
-            crate::config::Config::Compose(config) => config,
-            other => panic!("expected Compose config, got {other:?}"),
-        };
-        let compose = DockerCompose {
-            cli: ComposeCli::resolve().unwrap().0,
-            docker_host: Some("unix:///tmp/snouty-hermetic-test.sock".to_string()),
-        };
-
-        let output = compose.config_json_hermetic_env(&config).unwrap();
-        assert!(output.status.success(), "compose config failed: {output:?}");
-        let resolved: serde_json::Value = serde_json::from_slice(&output.stdout).unwrap();
-        let environment = &resolved["services"]["app"]["environment"];
-        assert_eq!(environment["HOME_VALUE"], "");
-        assert_eq!(environment["PATH_VALUE"], "");
-        assert_eq!(environment["DOCKER_HOST_VALUE"], "");
-    }
-
-    #[test]
-    fn compose_file_args_no_overlay() {
-        assert_eq!(
-            compose_file_args(None),
-            vec!["-f".to_string(), "docker-compose.yaml".to_string()]
-        );
-    }
-
-    #[test]
-    fn compose_file_args_with_overlay() {
-        let overlay = PathBuf::from("/tmp/override.yaml");
-        assert_eq!(
-            compose_file_args(Some(&overlay)),
-            vec![
-                "-f".to_string(),
-                "docker-compose.yaml".to_string(),
-                "-f".to_string(),
-                "/tmp/override.yaml".to_string(),
-            ]
         );
     }
